Rename `Money.currency` property to `metadata` and make static
Motivation:

The property name of `currency` was conflicting with the other usage of `Money` and `CurrencyMetadata`, so the new name is more accurate as to what the type info contains.

The information is also important to have both at runtime and compile time through statics and instance variables.

Modifications:

- Rename `Money.currency` to `Money.metadata`
- Change `Money.metadata` property to be static
- Add instance computed property of `metadata` to support runtime, existential, or instance checks of CurrencyMetadata

Result:

The name of properties should more accurate describe what is available with them, and there should be more flexibility for developers to access CurrencyMetadata