Add "sum" function for Sequences with AnyCurrency elements
Motivation:

It is common for developers to want a quick sum of values with currencies, to display totals.

To mirror functionality found on Sequence protocols with `max` and `min`, a `sum` with a predicate overload should
be provided.

Modifications:

- Add a `sum` function that starts with a zero value and adds all the elements together
- Add a `sum(where:)` overload that filters the elements by a predicate before summing

Result:

It should be more natural and easy for developers to quickly get a sum of currencies.