Refactor AnyCurrency storage for "minorUnits"
Motivation:

Currency algorithms are extremely prone to floating point precision variance when storing an `exactAmount`, when storing the raw `minorUnits` representation is more accurate and performant in most cases.

Currencies can be expressed a whole number representation in their minorUnits scale, and later converted for display and some arithmetic.

Modifications:

- Change `AnyCurrency` base value storage to be `Int64` of minor units instead of `Decimal` exact value
- Add `minorUnits` property
- Rename `init(minorUnits: Int)` to `init(exactly: Int64)`
- Add `init(exactly: Int64)` as a requirement for `AnyCurrency`
- Rename `roundedAmount` to `amount`
- Remove `exactAmount`
- Change arithmetic and comparison operators to use `amount` and `minorUnits` where appropriate

Result:

Developers should have direct access to the integer representation of a currency for arithmetic.