Add "XTS" currency
Motivation:

The "XTS" currency is an ISO 4217 specified currency for use in tests, so it stands to reason to explicitly support
this currency to allow developers to have a reasonable currency for tests.

As ISO 4217 does not specify any minor units for XTS, the minor unit will be `1` to be capable of validating rounding.

Modifications:

- Add: `XTS` currency definition

Result:

Developers should have access to the official ISO 4217 `XTS` test currency.