Change `XXX.minorUnits` to `0`
Motivation:

ISO 4217 leaves the 'minorUnits' or 'exponent' of several currencies as 'undefined'. For example, both XTS and XXX have no specified value.

XTS was given the value of '1' arbitrarily for the purposes of testing rounding behaviors, while XXX should be 0 as "no currency" is used
in transactions with XXX.

Modification:

- Change: `XXX.minorUnits` to be `0`

Result:

`XXX` currency values should now be whole number representations.