Refactor CurrencyMint
Motivation:

The current implementation of CurrencyMint is highly rigid and requires a ton of support from SwiftCurrency to be of true value.

Cases of custom currencies and quick fallbacks to a default type requires a fair bit of boilerplate code that can be simplified with
an improved re-implementation.

In addition, there were several places of code duplication that are possible to simplify with Swift's literal syntax for easier maintenance.

Modifications:

- Add: `CurrencyIdentifier` for CurrencyMint to use in lookup resolution methods
- Add: `standard` static singleton for CurrencyMint to resolve only ISO 4217 currencies
- Change: `CurrencyMint` to store a fallback lookup closure for 3rd party and default currency resolution

Result:

Developers using CurrencyMint should have an easier time supporting their own currencies, or unsupported/missing ISO 4217 currencies.