Replace CSV with JSON as file format for ISO Currency definitions
Motivation:

We're pulling in a dependency for CSV parsing as part of the build tool, forcing a transitive dependency on all users.

We simply need a way of storing the original definitions from the ISO Standard committee and parsing them as part of the build process, and JSON + `Codable` solves that for us easily.

Modifications:

- Change: Code to be a bit more legible on what's going on with the codegen
- Change: Data parsing from CSV to JSON using `Decodable`
- Rename: Files to be a bit more easily understandable of what's in each file
- Remove: **SwiftCSV** dependency

Result:

The build plugin is a bit easier to maintain, and there is now no dependencies in the project.