Re-implement the basic concept of 'Currency' with Decimal storage (#40)
Motivation:

Decimal provides the correct semantics for representing currency as it
guarantees representation of values exactly, as opposed to floating
point representations that may be incorrect. This also allows for
correct representations of fractional values during arithmetic, delaying
rounding until the user needs it to avoid "penny shaving".

In addition, the new implementation has much tighter semantics of how
its expected to use a `Currency` value, by not conforming to
`AdditiveArithmetic` and not providing the ability to multiply or divide
two `Currency` values

Modifications:
- Add: New plain protocol `Currency`
- Add: Tests for new protocols and extensions
- Change: Code generation of ISO currencies to include copies conforming
to the new `Currency` protocol
- Rename: `CurrencyMetadata` to `CurrencyDescriptor`
- Remove: `distributedProportionally` algorithm from `Currency`

Result:
The `Currency` type is more correct, with better ergonomics for usage in
a wider array of contexts, while avoiding common pitfalls.