Update CI test file to handle latest supported versions of Swift (#44)
Motivation:

The minimum version of our library is now Swift 5.7, and our CI does not properly reflect that. In addition, our CI file is manually defined, making it difficult to maintain as platforms come and go.

Modifications:

- Change: CI workflow to use `strategy: matrix` for defining jobs
- Change: CI workflow to cover Linux & macOS across Swift 5.7 - 5.10
- Change: Codecov workflow to make use of Vapor's re-useable GitHub Action

Result:

Maintainers of the library will have an easier time maintaining the platforms that tests run under as well as end-users will have greater confidence that the library will work as advertised on their desired platform.