Simplify the CurrencyValue protocol (#46)
Motivations:

The current `CurrencyValue` protocol has a few requirements with default
implementations that are not meant to be extension points on the
protocol.

In addition, a few aspects of how the protocol can be used are a bit
rough in either performance or generic ergonomics.

Modifications:

- Add: `AdditiveArithmetic` conformance for free
- Add: `@inlinable` annotations to remaining mutating and localization
methods
- Change: `minorUnits` and `roundedAmount` properties / initializers to
be pure extensions

Result:

Developers are able to use `CurrencyValue` in greater generic contexts
with slightly better performance capabilities, while also having more
protection from incorrectly override expected behavior.