Provide instance level access to descriptor information
Motivation:

Sometimes developers are working in type-erased contexts where they can't directly refer to the name of the type of the `CurrencyValue` and need access to the descriptor.

In this cases, having an instance member property to provide the value is straight forward and convenient.

Modifications:

- Add: `descriptor` instance property to `CurrencyValue`
- Add: `@inlinable` annotations to both the static and instance member `descriptor` properties

Result:

Developers now have access to the `CurrencyValue` descriptor type either at the type or instance level.