Update project documentation for new state of the world (#49)
Motivation:

A hallmark of a good library is its documentation, and much of the
library was re-written. The documentation needs to be updated to avoid
giving incorrect information to developers, but also can be made more
robust by going in depth on certain topics.

In addition, the README needed to be updated for several broken links
and new infrastructure in use such as for GitHub Actions and Swift
Package Index documentation hosting.

Modifications:

- Add: Unit tests of code samples used in documentation
- Change: README to have updated links and information
- Change: Symbol documentation and articles significantly to be more
in-depth, professional, and up to date
- Remove: No longer used typealias for `CurrencyDescriptor`

Result:

Developers looking to learn about the library and its concepts should
have a more straight forward and clear understanding from within our
code base itself.