Add explicit support for Swift 6
Motivation:

Much of the ecosystem is starting to make the migration to Swift 6. Some
even build with warnings as errors.

The current project has many deprecation warnings when compiling as
Swift 6.

Modifications:

- Add: Swift 5.7 backwards support Package.swift
- Change: Usage of #file in tests to #filePath
- Change: Implementation of Codegen plugin to not use deprecated
properties and methods

Result:

Existing users should not see any build changes, while those building
with Swift 6 should be able to continue using the project with no issues